Move statsd handling to observability module
