+++ src/observability.py
@@ -3,6 +3,8 @@
 
 """Provide the Observability class to represent the observability stack for Flask application."""
 
+import textwrap
+
 import ops
 from charms.grafana_k8s.v0.grafana_dashboard import GrafanaDashboardProvider
 from charms.loki_k8s.v0.loki_push_api import LogProxyConsumer
@@ -12,7 +14,7 @@ from charm_state import CharmState
 from constants import FLASK_CONTAINER_NAME
 
 
-class Observability:  # pylint: disable=too-few-public-methods
+class Observability(ops.Object):  # pylint: disable=too-few-public-methods
     """A class representing the observability stack for Flask application."""
 
     def __init__(self, charm: ops.CharmBase, charm_state: CharmState):
@@ -22,6 +24,8 @@ class Observability:  # pylint: disable=too-few-public-methods
             charm: The charm object that the Observability instance belongs to.
             charm_state: The state of the charm that the Observability instance belongs to.
         """
+        super().__init__(charm, "observability")
+        self._charm = charm
         self._metrics_endpoint = MetricsEndpointProvider(
             charm,
             relation_name="metrics-endpoint",
@@ -36,3 +40,49 @@ class Observability:  # pylint: disable=too-few-public-methods
         self._grafana_dashboards = GrafanaDashboardProvider(
             charm, relation_name="grafana-dashboard"
         )
+        self._charm.framework.observe(
+            self._charm.on.statsd_prometheus_exporter_pebble_ready,
+            self._on_statsd_prometheus_exporter_pebble_ready,
+        )
+
+    def _on_statsd_prometheus_exporter_pebble_ready(self, _event: ops.PebbleReadyEvent) -> None:
+        """Handle the statsd-prometheus-exporter-pebble-ready event."""
+        container = self._charm.unit.get_container("statsd-prometheus-exporter")
+        container.push(
+            "/statsd.conf",
+            textwrap.dedent(
+                """\
+                mappings:
+                  - match: gunicorn.request.status.*
+                    name: flask_response_code
+                    labels:
+                      status: $1
+                  - match: gunicorn.requests
+                    name: flask_requests
+                  - match: gunicorn.request.duration
+                    name: flask_request_duration
+                """
+            ),
+        )
+        statsd_layer = ops.pebble.LayerDict(
+            summary="statsd exporter layer",
+            description="statsd exporter layer",
+            services={
+                "statsd-prometheus-exporter": {
+                    "override": "replace",
+                    "summary": "statsd exporter service",
+                    "user": "nobody",
+                    "command": "/bin/statsd_exporter --statsd.mapping-config=/statsd.conf",
+                    "startup": "enabled",
+                }
+            },
+            checks={
+                "container-ready": {
+                    "override": "replace",
+                    "level": "ready",
+                    "http": {"url": "http://localhost:9102/metrics"},
+                },
+            },
+        )
+        container.add_layer("statsd-prometheus-exporter", statsd_layer, combine=True)
+        container.replan()
